requirements.txt
finance-datareader
pandas
tabulate
yfinance<1.0.0
requests
google-generativeai
beautifulsoup4
python-dotenv
